requirements.txt

pyyaml
ete3
numpy >= 1.13
tensorflow >= 1.4
pandas >= 0.20.3
biopython >= 1.70
scipy >= 0.19.0
khmer >= 2.1.1
pyfaidx >= 0.5
scikit-learn>=0.19.0
